Sketch: Create new flocker.control with corresponding flocker-control script, in charge of managing the cluster.
