+++ flocker/control/__init__.py
@@ -1,0 +1,5 @@
+# Copyright Hybrid Logic Ltd.  See LICENSE file for details.
+
+"""
+Control service for managing a Flocker cluster.
+"""
