+++ flocker/node/script.py
@@ -9,9 +9,6 @@ tools.
 import sys
 
 from twisted.python.usage import Options, UsageError
-from twisted.internet.defer import Deferred, maybeDeferred
-from twisted.internet.endpoints import TCP4ServerEndpoint
-from twisted.application.service import MultiService
 
 
 from yaml import safe_load, safe_dump
@@ -23,17 +20,17 @@ from ._config import marshal_configuration
 
 from ..volume.service import (
     ICommandLineVolumeScript, VolumeScript)
-from ..volume.httpapi import create_api_service
+
 from ..volume.script import flocker_volume_options
 from ..common.script import (
-    flocker_standard_options, FlockerScriptRunner)
+    flocker_standard_options, FlockerScriptRunner, main_for_service)
 from . import (ConfigurationError, model_from_configuration, Deployer,
                FlockerConfiguration, current_from_configuration)
 
 __all__ = [
     "flocker_changestate_main",
     "flocker_reportstate_main",
-    "flocker_serve_main",
+    "flocker_volume_main",
 ]
 
 
@@ -211,87 +208,12 @@ def flocker_reportstate_main():
     ).main()
 
 
-def _chain_stop_result(service, stop):
-    """
-    Stop a service and chain the resulting ``Deferred`` to another
-    ``Deferred``.
-
-    :param IService service: The service to stop.
-    :param Deferred stop: The ``Deferred`` which will be fired when the service
-        has stopped.
-    """
-    maybeDeferred(service.stopService).chainDeferred(stop)
-
-
-def _main_for_service(reactor, service):
-    """
-    Start a service and integrate its shutdown with reactor shutdown.
-
-    This is useful for hooking driving an ``IService`` provider with
-    ``twisted.internet.task.react``.  For example::
-
-        from twisted.internet.task import react
-        from yourapp import YourService
-        react(_main_for_service, [YourService()])
-
-    :param IReactorCore reactor: The reactor the run lifetime of which to tie
-        to the given service.  When the reactor is shutdown, the service will
-        be shutdown.
-
-    :param IService service: The service to tie to the run lifetime of the
-        given reactor.  It will be started immediately and made to stop when
-        the reactor stops.
-
-    :return: A ``Deferred`` which fires after the service has finished
-        stopping.
-    """
-    service.startService()
-    stop = Deferred()
-    reactor.addSystemEventTrigger(
-        "before", "shutdown", _chain_stop_result, service, stop)
-    return stop
-
-
 @flocker_standard_options
 @flocker_volume_options
 class ServeOptions(Options):
     """
     Command line options for ``flocker-serve`` cluster management process.
     """
-    optParameters = [
-        ["port", "p", 4523, "The port to listen on.", int],
-        ]
-
-
-class _ServeService(MultiService):
-    """
-    Service for running a ``VolumeService`` and HTTP API service.
-    """
-    def __init__(self, volume_service, http_service):
-        """
-        :param volume_service: The volume service to run.
-
-        :param http_service: The HTTP API service to run.
-        """
-        MultiService.__init__(self)
-        volume_service.setServiceParent(self)
-        http_service.setServiceParent(self)
-
-    def stopService(self):
-        """
-        Stop all services, return result of stopping the volume service.
-
-        The volume service is the service whose failure during stopping
-        matters (and therefore should result in non-zero exit code) since
-        it is the one likely to have significant code. By default a
-        MultiService returns a ``DeferredList``, though, so errors in
-        stopping the volume service will be swallowed.
-
-        :return: Result from stopping the volume service.
-        """
-        d = MultiService.stopService(self)
-        d.addCallback(lambda results: results[-1][1])
-        return d
 
 
 @implementer(ICommandLineVolumeScript)
@@ -301,13 +223,10 @@ class ServeScript(object):
     a Flocker cluster.
     """
     def main(self, reactor, options, volume_service):
-        api_service = create_api_service(
-            TCP4ServerEndpoint(reactor, options["port"]))
-        parent_service = _ServeService(volume_service, api_service)
-        return _main_for_service(reactor, parent_service)
+        return main_for_service(reactor, volume_service)
 
 
-def flocker_serve_main():
+def flocker_volume_main():
     return FlockerScriptRunner(
         script=VolumeScript(ServeScript()),
         options=ServeOptions()
